disable tests when building container
Might be nice if there is a blazingly fast direct internet connection.
Takes half an hour when behind a slow proxy and then just fails
because the proxy didn't respond in time for one of the tests.

The installation doesn't even stop when there is an error,
it just goes on and then tells you it didn't work in the end.